requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
community==1.0.0b1
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0
Flask==3.1.1
fonttools==4.58.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langsmith==0.3.43
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
numpy==2.2.6
openai==1.82.1
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.0
zstandard==0.23.0
